+++ philo_one/Includes/utils.h
@@ -4,7 +4,7 @@
 # include <string.h>
 # include <unistd.h>
 
-int			ft_atoi(const char *str);
+long int	ft_atoi(const char *str);
 size_t		ft_strlen(const char *s);
 void		ft_putstr_fd(char *s, int fd);
 int			ft_isdigit(int c);
